Allow a user to configure audio type to skip detection
This should make swob a bit faster for eg. ALSA users (without pipewire nor
pulse), at the cost of it checking config directories all the time rather
than only when it starts wob. Really the check should be moved to the daemon
side, the daemon should be a shellscript which calls set_vol and feeds it
to wob, and the controller should just echo "$*" >$wobfifo. That's an idea